Rename `droplet` backend to `doDroplet`
In order to maintain a clear namespace for all attributes in the plugin,
we add a `do` prefix to the `droplet` backend. This would ensure
resources would not be confused between plugins; a good example is
a future `volume` resource, which may end up being confusing in the event
multiple cloud backends are used in a single project.